account type guard moved
